plumbing: http, use buffer HTTP I/O
This uses io.TeeReader to read the request body and write it to the
response body as it is read. This allows the server to send the request
body to the client as it is being read, instead of waiting for the
entire request body to be read before sending the response.

This also implements a custom flushResponseWriter that handles buffered
output and implements the io.ReaderFrom interface. This allows the
server to write the response body in chunks and flush it to the client
as it is being written to avoid large memory usage.